remove useless code, clean code, ...
