+++ package.json
@@ -4,7 +4,7 @@
   "license": "MIT",
   "main": "src/index.js",
   "dependencies": {
-    "@ant-design/icons": "4.1.0",
+    "@ant-design/icons": "^4.1.0",
     "@apollo/react-hooks": "0.1.0-beta.11",
     "antd": "4.2.0",
     "apollo-boost": "0.4.3",
@@ -14,6 +14,7 @@
     "apollo-link-http": "^1.5.17",
     "graphql": "latest",
     "graphql-tag": "latest",
+    "history": "^4.10.1",
     "jquery": "^3.5.1",
     "lodash": "4.17.15",
     "react": "16.8.6",
