+++ src/app.js
@@ -34,7 +34,7 @@ class App extends Component {
     this.setState({
       collapsed: !this.state.collapsed
     });
-  };
+  };  
 
   render() {
     return (
