+++ src/index.js
@@ -21,7 +21,7 @@ const httpLink = createHttpLink({
 
 const authLink = setContext((_, { headers }) => {
   // const token = localStorage.getItem(AUTH_TOKEN);
-
+  // window.localStorage.clear();
   return {
     headers: {
       ...headers
