--- src/pages/questionnaries/quiz.json
@@ -1,7 +1,0 @@
-{
-    "id":1,
-    "introduction":"Here, you will practice.",
-    "numOfQuestions":0,
-    "date":20200520,
-    "questions":null
-}
